servers/mcp/requirements.txt
fastmcp
google-adk
python-dotenv
uvicorn
